move view into custom op, update bfs with bound, update custom op return type to be half precision

+++ tests/unittest/_torch/auto_deploy/unit/singlegpu/transformations/library/test_rope_matcher.py
@@ -56,7 +56,7 @@ class RotaryModel(torch.nn.Module):
         sin = sin.to(q.device)
 
         q_embed, k_embed = apply_rotary_pos_emb(q, k, cos, sin, unsqueeze_dim=0)
-        return q_embed + k_embed
+        return (q_embed + k_embed).to(torch.float16)
 
 
 @torch.inference_mode()
